Swap telemetry backend from Supabase to Cloudflare Workers + D1
