+++ openspec/changes/add-opt-in-telemetry/proposal.md
@@ -6,14 +6,14 @@ The audience is self-hosted Jellyfin users — the most telemetry-hostile popula
 
 ## What Changes
 
-- New Supabase backend (free tier): one `pings` table, one `/ingest` edge function (validates shape, caps size, rate-limits, never persists IPs into the dataset), RLS deny-all so the shipped anon key cannot touch the table directly.
+- New Cloudflare Workers + D1 backend (free tier, never idle-pauses): one `pings` table, one ingest Worker (validates shape + key, caps size, rate-limits, never persists IPs into the dataset); the database has no client-facing access path at all.
 - New plugin config: `TelemetryEnabled` (default false), `TelemetryInstanceId` (random UUID generated on enable, user-regenerable), persisted window counters and per-category error state.
 - New `TelemetryTask : IScheduledTask`: weekly anonymous ping with ±6 h jitter; payload is plugin/Jellyfin version, feature-toggle booleans, bucketed counts (never raw numbers), and per-period error-category counts.
 - Error-transition pings: when any error category goes clean → failing, a ping carrying the full error-state map fires, capped at one per instance per day (further transitions queue and consolidate).
 - Settings UI: opt-in checkbox, UUID regenerate button, admin-only "Preview exact JSON" modal that doubles as a copy-diagnostic-bundle for bug reports (with an explicit deanonymization warning).
 - One-time dismissible opt-in banner on the plugin dashboard tab, shipped as a separate release after the pipe is proven.
 - README "Telemetry" section documenting the exact payload and the precisely-worded anonymity promise.
-- Two scheduled GitHub Actions: `telemetry-keepalive.yml` (daily DB touch so the free tier never pauses; self-sustaining against GitHub's 60-day cron auto-disable) and `telemetry-canary.yml` (daily regression check that auto-files gated, deduped GitHub issues).
+- One scheduled GitHub Action: `telemetry-canary.yml` (daily regression check via wrangler/D1 that auto-files gated, deduped GitHub issues, doubles as the abuse alarm, and keeps its own cron alive against GitHub's 60-day auto-disable).
 
 ## Capabilities
 
@@ -28,8 +28,8 @@ The audience is self-hosted Jellyfin users — the most telemetry-hostile popula
 ## Impact
 
 - **New plugin code**: `TelemetryTask`, payload builder, error-state tracking wired into existing failure paths, `GET /Telemetry/Preview` admin endpoint, config-page banner/checkbox/modal. Two plugin releases (pipe first, prompt second), both through the normal version-gated pipeline.
-- **New infrastructure**: one Supabase project (migration + edge function), maintained near-zero; the maintainer queries it via the existing Supabase MCP connection.
-- **New workflows**: `telemetry-keepalive.yml`, `telemetry-canary.yml`; repo secrets gain the Supabase read key. The keep-alive heartbeat commits only to a dedicated non-main branch so the every-merge-ships release pipeline is never triggered by it.
+- **New infrastructure**: one Cloudflare Worker + D1 database (worker/ in the repo: wrangler.toml, schema.sql, src/index.ts), maintained near-zero; the maintainer queries it via wrangler / the D1 HTTP API with a scoped token.
+- **New workflow**: `telemetry-canary.yml`; repo secrets gain TELEMETRY_CF_API_TOKEN and TELEMETRY_CF_ACCOUNT_ID. Its heartbeat commits only to a dedicated non-main branch so the every-merge-ships release pipeline is never triggered by it.
 - **README** gains the telemetry section; the privacy wording is load-bearing and reviewed like code.
 - **No behaviour change for users who don't opt in**: telemetry is dead code until the checkbox is ticked; the only visible change is the one-time banner.
 - **Shipped write key is extractable by design**: worst case is junk rows (data-quality problem bounded by validation, rate limits, and a growth alarm), never a privacy problem.
